Moving some things around because its easier that way. Adding note to readme about compiling it with local toolchain.
